Remove counting logic from BaseComparison and put it only in GTcomparison
