+++ spiketoolkit/comparison/basecomparison.py
@@ -1,7 +1,7 @@
 import spikeextractors as se
 
 from .comparisontools import (count_matching_events, compute_agreement_score,
-                                                do_matching, do_score_labels, do_counting, do_confusion_matrix)
+                                                do_matching, do_score_labels,  do_confusion_matrix)
 
 
 class BaseComparison:
@@ -10,7 +10,7 @@ class BaseComparison:
     
     """
     def __init__(self, sorting1, sorting2, sorting1_name=None, sorting2_name=None, delta_frames=10, min_accuracy=0.5,
-                 count=False, n_jobs=1, verbose=False):
+                n_jobs=1, verbose=False):
         self._sorting1 = sorting1
         self._sorting2 = sorting2
         self.sorting1_name = sorting1_name
@@ -18,15 +18,20 @@ class BaseComparison:
         self._delta_frames = delta_frames
         self._min_accuracy = min_accuracy
         self._n_jobs = n_jobs
+        
+        # maching is done always
         if verbose:
             print("Matching...")
         self._do_matching()
+        
+        # make score label is done always
+        # Samuel EDIT : should do_score_label done on demand ?
+        # I think no.
+        self._do_score_labels()
+        
+        # confusion matrix is compute on demand
+        self._confusion_matrix = None
 
-        self._mixed_counts = None
-        if count:
-            if verbose:
-                print("Counting...")
-            self._do_counting(verbose=verbose)
 
     def get_sorting1(self):
         # Samuel EDIT : why not a direct attribute acees  with self.sorting1 ?
@@ -49,18 +54,14 @@ class BaseComparison:
             raise Exception("Unit_id is not a valid unit")
 
 
-    def compute_counts(self):
-        if self._mixed_counts is None:
-            self._do_counting(verbose=False)
-
 
     def plot_confusion_matrix(self, xlabel=None, ylabel=None):
         # Samuel EDIT
         # This must be moved in spikewidget
         import matplotlib.pylab as plt
-
-        if self._mixed_counts is None:
-            self._do_counting(verbose=False)
+        
+        if self._confusion_matrix is None:
+            self._do_confusion()
 
         sorting1 = self._sorting1
         sorting2 = self._sorting2
@@ -68,7 +69,7 @@ class BaseComparison:
         unit2_ids = sorting2.get_unit_ids()
         N1 = len(unit1_ids)
         N2 = len(unit2_ids)
-        st1_idxs, st2_idxs = self._do_confusion()
+        
         fig, ax = plt.subplots()
         # Using matshow here just because it sets the ticks up nicely. imshow is faster.
         ax.matshow(self._confusion_matrix, cmap='Greens')
@@ -88,8 +89,8 @@ class BaseComparison:
         ax.set_yticks(np.arange(0, N1 + 1))
         ax.xaxis.tick_bottom()
         # Labels for major ticks
-        ax.set_xticklabels(np.append(st2_idxs, 'FN'), fontsize=12)
-        ax.set_yticklabels(np.append(st1_idxs, 'FP'), fontsize=12)
+        ax.set_xticklabels(np.append(self._st2_idxs, 'FN'), fontsize=12)
+        ax.set_yticklabels(np.append(self._st1_idxs, 'FP'), fontsize=12)
 
         if xlabel == None:
             if self.sorting2_name is None:
@@ -107,23 +108,20 @@ class BaseComparison:
             ax.set_ylabel(ylabel, fontsize=20)
 
         return ax
+
     def _do_matching(self):
         self._event_counts_1,  self._event_counts_2, self._matching_event_counts_12,\
             self._best_match_units_12, self._matching_event_counts_21,\
             self._best_match_units_21,self._unit_map12,\
             self._unit_map21 = do_matching(self._sorting1, self._sorting2, self._delta_frames, self._min_accuracy, self._n_jobs)
 
-    def _do_counting(self, verbose=False):
+    def _do_score_labels(self, verbose=False):
         self._labels_st1, self._labels_st2 = do_score_labels(self._sorting1, self._sorting2,
                                                              self._delta_frames, self._unit_map12)
-        self._mixed_counts = do_counting(self._sorting1, self._sorting2, self._unit_map12,
-                                         self._labels_st1, self._labels_st2)
 
     def _do_confusion(self):
-        self._confusion_matrix,  st1_idxs, st2_idxs = do_confusion_matrix(self._sorting1, self._sorting2,
+        self._confusion_matrix,  self._st1_idxs, self._st2_idxs = do_confusion_matrix(self._sorting1, self._sorting2,
                                                 self._unit_map12, self._labels_st1, self._labels_st2)
 
-        return st1_idxs, st2_idxs
-
