+++ spiketoolkit/comparison/tests/test_sortingcomparison.py
@@ -21,7 +21,7 @@ def test_compare_two_sorters():
     # simple match
     sorting1, sorting2 = make_sorting([100, 200, 300, 400], [0, 0, 1, 0], 
                                                             [101, 201, 301, ], [0, 0, 5])
-    sc = compare_two_sorters(sorting1, sorting2, count=True)
+    sc = compare_two_sorters(sorting1, sorting2)
     
     sc._do_confusion()
     print(sc._confusion_matrix)
